ezqms-762: extract base content from toc popup Signed-off-by: Alexey Zinoviev <alexey.zinoviev@xored.com>
